+++ monitor-e2e/test/ercToErc.js
@@ -1,23 +1,9 @@
 const assert = require('assert')
 const axios = require('axios')
-const Web3 = require('web3')
-const shell = require('shelljs')
 const { ercToErcBridge, user, homeRPC, foreignRPC } = require('../../e2e-commons/constants.json')
+const { checkAll, sendEther } = require('../utils')
 
 const baseUrl = ercToErcBridge.monitor
-const HOME_BRIDGE_ADDRESS = ercToErcBridge.home
-const FOREIGN_BRIDGE_ADDRESS = ercToErcBridge.foreign
-const homeWeb3 = new Web3(new Web3.providers.HttpProvider(homeRPC.URL))
-const foreignWeb3 = new Web3(new Web3.providers.HttpProvider(foreignRPC.URL))
-homeWeb3.eth.accounts.wallet.add(user.privateKey)
-foreignWeb3.eth.accounts.wallet.add(user.privateKey)
-
-const checkAll = () => {
-  // the different container names comes from macOS/linux docker differences and/or different docker-compose versions
-  ;['e2e-commons_monitor_1', 'e2e_commons_monitor_1', 'e2ecommons_monitor_1'].forEach(container =>
-    shell.exec(`docker exec ${container} yarn check-all`)
-  )
-}
 
 describe('ERC TO ERC', () => {
   let data
@@ -30,8 +16,7 @@ describe('ERC TO ERC', () => {
   it('should contain totalSupply', () => assert(data.home.totalSupply === '0'))
 })
 
-
-describe.only('ERC TO ERC with changing state of contracts', () => {
+describe('ERC TO ERC with changing state of contracts', () => {
   let data
 
   before(async () => {
@@ -39,52 +24,18 @@ describe.only('ERC TO ERC with changing state of contracts', () => {
   })
 
   it('should change balanceDiff', async () => {
-    console.log(user.address, HOME_BRIDGE_ADDRESS)
-    // send transaction to home chain
-    const depositTx = await homeWeb3.eth.sendTransaction({
-      from: user.address,
-      to: HOME_BRIDGE_ADDRESS,
-      gasPrice: '1',
-      gas: '50000',
-      value: '1000000000000000000'
-    })
-
-    // Send a trivial transaction to generate a new block since the watcher
-    // is configured to wait 1 confirmation block
-    // await generateNewBlock(homeWeb3, user.address)
-
-    console.log('first should be zero...')
-    // assert(data.balanceDiff === 0)
-
+    assert(data.balanceDiff === 0)
+    await sendEther(homeRPC.URL, user, ercToErcBridge.home)
     checkAll()
     ;({ data } = await axios.get(`${baseUrl}`))
-
-    console.log('should not be zero...')
     assert(data.balanceDiff !== 0)
   })
 
   it('should change balanceDiff', async () => {
-    console.log(user.address, HOME_BRIDGE_ADDRESS)
-    // send transaction to home chain
-    const depositTx = await foreignWeb3.eth.sendTransaction({
-      from: user.address,
-      to: FOREIGN_BRIDGE_ADDRESS,
-      gasPrice: '1',
-      gas: '50000',
-      value: '1000000000000000000'
-    })
-
-    // Send a trivial transaction to generate a new block since the watcher
-    // is configured to wait 1 confirmation block
-    // await generateNewBlock(homeWeb3, user.address)
-
-    console.log('first should be zero...')
-    // assert(data.balanceDiff === 0)
-
+    assert(data.balanceDiff === 0)
+    await sendEther(foreignRPC.URL, user, ercToErcBridge.foreign)
     checkAll()
     ;({ data } = await axios.get(`${baseUrl}`))
-
-    console.log('should not be zero...')
     assert(data.balanceDiff !== 0)
   })
 })
